Fix to 'repo-svn push' command
Trying to push an existing podspec with an increased version number
fails to update the repository.

Issue fails with:

svn: warning: W150002:
'/Users/user/.cocoapods/repos/PrivateRepo/MyProject’ is already under
version control

The solution is to add the —force command to the ‘svn add’ operation,
and the commit happens successfully.